requirements.txt
fastapi
uvicorn
tortoise-orm
cryptography
websockets
pyjwt
